gemini-folder-organizer/requirements.txt
google-generativeai
python-docx
openpyxl
pymupdf
pillow
python-dotenv
pydantic
